Safari-611.1.10/LayoutTests/editing/inserting/insert-3907422-fix.html

<html> 
<head>

<style>
.editing { 
    word-wrap: break-word; 
    -khtml-nbsp-mode: space; 
    -khtml-line-break: after-white-space;
}
</style>

<title>Editing Test</title> 
</head> 
<body contenteditable id="test" class="editing">
<div>foo<blockquote>bar</blockquote>baz</div>

<script src="../../resources/dump-as-markup.js"></script>
<script>
    document.execCommand("SelectAll");
    document.execCommand("Cut");
    document.execCommand("Paste");
    document.execCommand("Paste");
    Markup.description("Buggy code before fix would insert blockquote after body element in second paste");
    Markup.dump();
</script>

</body>
</html>
